requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
_tflow_select=2.3.0=mkl
abseil-cpp=20211102.0=hd4dd3e8_0
absl-py=1.4.0=pypi_0
aiohttp=3.8.3=py38h5eee18b_0
aiosignal=1.2.0=pyhd3eb1b0_0
appdirs=1.4.4=pyhd3eb1b0_0
asttokens=2.2.1=pypi_0
astunparse=1.6.3=py_0
async-timeout=4.0.2=py38h06a4308_0
attrs=22.1.0=py38h06a4308_0
autopep8=1.6.0=pyhd3eb1b0_1
backcall=0.2.0=pypi_0
blas=1.0=mkl
blinker=1.4=py38h06a4308_0
brotlipy=0.7.0=py38h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
c-ares=1.19.0=h5eee18b_0
ca-certificates=2023.01.10=h06a4308_0
cachetools=5.3.0=pypi_0
certifi=2022.12.7=py38h06a4308_0
cffi=1.15.1=py38h5eee18b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py38h06a4308_0
cloudpickle=2.2.1=pypi_0
contourpy=1.0.7=pypi_0
cryptography=38.0.4=py38h9ce1e76_0
cuda=11.7.1=0
cuda-cccl=11.7.91=0
cuda-command-line-tools=11.7.1=0
cuda-compiler=11.7.1=0
cuda-cudart=11.7.99=0
cuda-cudart-dev=11.7.99=0
cuda-cuobjdump=11.7.91=0
cuda-cupti=11.7.101=0
cuda-cuxxfilt=11.7.91=0
cuda-demo-suite=12.0.140=0
cuda-documentation=12.0.140=0
cuda-driver-dev=11.7.99=0
cuda-gdb=12.0.140=0
cuda-libraries=11.7.1=0
cuda-libraries-dev=11.7.1=0
cuda-memcheck=11.8.86=0
cuda-nsight=12.0.140=0
cuda-nsight-compute=12.0.1=0
cuda-nvcc=11.7.99=0
cuda-nvdisasm=12.0.140=0
cuda-nvml-dev=11.7.91=0
cuda-nvprof=12.0.146=0
cuda-nvprune=11.7.91=0
cuda-nvrtc=11.7.99=0
cuda-nvrtc-dev=11.7.99=0
cuda-nvtx=11.7.91=0
cuda-nvvp=12.0.146=0
cuda-runtime=11.7.1=0
cuda-sanitizer-api=12.0.140=0
cuda-toolkit=11.7.1=0
cuda-tools=11.7.1=0
cuda-visual-tools=11.7.1=0
cycler=0.11.0=pypi_0
decorator=5.1.1=pypi_0
docopt=0.6.2=pypi_0
executing=1.2.0=pypi_0
farama-notifications=0.0.4=pypi_0
ffmpeg=1.4=pypi_0
ffmpeg-python=0.2.0=pypi_0
filelock=3.9.0=pypi_0
flatbuffers=23.3.3=pypi_0
flit-core=3.6.0=pyhd3eb1b0_0
fonttools=4.38.0=pypi_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.3.3=py38h5eee18b_0
future=0.18.3=pypi_0
gast=0.4.0=pyhd3eb1b0_0
gds-tools=1.5.1.14=0
giflib=5.2.1=h5eee18b_1
gitdb=4.0.10=pypi_0
gitpython=3.1.31=pypi_0
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
google-auth=2.16.1=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-pasta=0.2.0=pyhd3eb1b0_0
grpc-cpp=1.46.1=h33aed49_1
grpcio=1.51.1=pypi_0
gviz-api=1.10.0=pypi_0
gym=0.26.2=pypi_0
gym-notices=0.0.8=pypi_0
gymnasium=0.28.1=pypi_0
h5py=3.8.0=pypi_0
hdf5=1.10.6=h3ffc7dd_1
huggingface-hub=0.12.1=pypi_0
icu=58.2=he6710b0_3
idna=3.4=py38h06a4308_0
imageio=2.26.0=pypi_0
importlib-metadata=6.0.0=py38h06a4308_0
importlib-resources=5.12.0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
ipython=8.10.0=pypi_0
jax=0.4.7=pypi_0
jax-jumpy=1.0.0=pypi_0
jedi=0.18.2=pypi_0
jpeg=9e=h7f8727e_0
keras=2.12.0=pypi_0
keras-preprocessing=1.1.2=pyhd3eb1b0_0
kiwisolver=1.4.4=pypi_0
krb5=1.19.4=h568e23c_0
lame=3.100=h7b6447c_0
lazy-loader=0.1=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libclang=16.0.0=pypi_0
libcublas=11.10.3.66=0
libcublas-dev=11.10.3.66=0
libcufft=10.7.2.124=h4fbf590_0
libcufft-dev=10.7.2.124=h98a8f43_0
libcufile=1.5.1.14=0
libcufile-dev=1.5.1.14=0
libcurand=10.3.1.124=0
libcurand-dev=10.3.1.124=0
libcurl=7.88.1=h91b91d3_0
libcusolver=11.4.0.1=0
libcusolver-dev=11.4.0.1=0
libcusparse=11.7.4.91=0
libcusparse-dev=11.7.4.91=0
libdeflate=1.8=h7f8727e_5
libedit=3.1.20221030=h5eee18b_0
libev=4.33=h7f8727e_1
libffi=3.4.2=h6a678d5_6
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.2=h7f8727e_0
libnghttp2=1.46.0=hce63b2e_0
libnpp=11.7.4.75=0
libnpp-dev=11.7.4.75=0
libnvjpeg=11.8.0.2=0
libnvjpeg-dev=11.8.0.2=0
libpng=1.6.37=hbc83047_0
libprotobuf=3.20.3=he621ea3_0
libssh2=1.10.0=h8f2d780_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.16.0=h27cfd23_0
libtiff=4.5.0=h6a678d5_1
libunistring=0.9.10=h27cfd23_0
libwebp=1.2.4=h11a3e52_0
libwebp-base=1.2.4=h5eee18b_0
llvmlite=0.39.1=pypi_0
lz4-c=1.9.4=h6a678d5_0
markdown=3.4.1=py38h06a4308_0
markupsafe=2.1.2=pypi_0
matplotlib=3.7.0=pypi_0
matplotlib-inline=0.1.6=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
ml-dtypes=0.0.4=pypi_0
multidict=6.0.2=py38h5eee18b_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=3.0=pypi_0
nsight-compute=2022.4.1.6=0
numba=0.56.4=pypi_0
numpy=1.23.5=py38h14f4228_0
numpy-base=1.23.5=py38h31eccc5_0
oauthlib=3.2.2=py38h06a4308_0
object-detection-metrics=0.4.post1=pypi_0
opencv-python=4.7.0.68=pypi_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1t=h7f8727e_0
opt_einsum=3.3.0=pyhd3eb1b0_1
packaging=23.0=py38h06a4308_0
pandas=1.5.3=pypi_0
parso=0.8.3=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.3.0=py38h6a678d5_2
pip=22.3.1=py38h06a4308_0
pooch=1.4.0=pyhd3eb1b0_0
prompt-toolkit=3.0.36=pypi_0
protobuf=4.22.0=pypi_0
psutil=5.9.4=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=pypi_0
pycocotools=2.0.6=pypi_0
pycodestyle=2.10.0=py38h06a4308_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.14.0=pypi_0
pyjwt=2.4.0=py38h06a4308_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=pypi_0
pysocks=1.7.1=py38h06a4308_0
python=3.8.16=h7a1cb2a_2
python-dateutil=2.8.2=pypi_0
python-flatbuffers=2.0=pyhd3eb1b0_0
python-version=0.0.2=pypi_0
pytorch=1.13.1=py3.8_cuda11.7_cudnn8.5.0_0
pytorch-cuda=11.7=h67b0de4_1
pytorch-mutex=1.0=cuda
pytz=2022.7.1=pypi_0
pywavelets=1.4.1=pypi_0
pyyaml=6.0=pypi_0
re2=2022.04.01=h295c915_0
readline=8.2=h5eee18b_0
regex=2022.10.31=pypi_0
requests=2.28.1=py38h06a4308_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
scikit-image=0.20.0=pypi_0
scipy=1.9.1=pypi_0
seaborn=0.12.2=pypi_0
setuptools=65.6.3=py38h06a4308_0
shapely=2.0.1=pypi_0
six=1.16.0=pyhd3eb1b0_1
smmap=5.0.0=pypi_0
snappy=1.1.9=h295c915_0
sqlite=3.40.1=h5082296_0
stack-data=0.6.2=pypi_0
tensorboard=2.12.0=pypi_0
tensorboard-data-server=0.7.0=pypi_0
tensorboard-plugin-profile=2.11.2=pypi_0
tensorboard-plugin-wit=1.8.1=py38h06a4308_0
tensorflow=2.10.0=mkl_py38hd2379f1_0
tensorflow-base=2.10.0=mkl_py38hb9daa73_0
tensorflow-estimator=2.12.0=pypi_0
tensorflow-io-gcs-filesystem=0.31.0=pypi_0
termcolor=2.2.0=pypi_0
thop=0.1.1-2209072238=pypi_0
tianshou=0.5.0=pypi_0
tifffile=2023.2.28=pypi_0
timm=0.6.12=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.13.2=pypi_0
toml=0.10.2=pyhd3eb1b0_0
torch-tb-profiler=0.4.1=pypi_0
torchaudio=0.13.1=py38_cu117
torchvision=0.14.1=py38_cu117
tqdm=4.64.1=pypi_0
traitlets=5.9.0=pypi_0
transformers=4.26.1=pypi_0
typing_extensions=4.4.0=py38h06a4308_0
urllib3=1.26.14=py38h06a4308_0
wcwidth=0.2.6=pypi_0
werkzeug=2.2.3=py38h06a4308_0
wheel=0.38.4=py38h06a4308_0
wrapt=1.14.1=py38h5eee18b_0
xz=5.2.10=h5eee18b_1
yarl=1.8.1=py38h5eee18b_0
zipp=3.14.0=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.2=ha4553b6_0